Assignment1/google_news/assignment/requirements.txt
'django-cors-headers' 
djangorestframework_camel_case 
